+++ commerce/commerce/__init__.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import, unicode_literals
 # This will make sure the app is always imported when
 # Django starts so that shared_task will use this app.
 from commerce.celery import app as celery_app
