+++ commerce/commerce/celery.py
@@ -1,6 +1,7 @@
+from __future__ import absolute_import, unicode_literals
+
 import os
 from celery import Celery
-from commerce.settings import INSTALLED_APPS
 
 # set the default Django settings module for the 'celery' program.
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'commerce.settings')
@@ -14,5 +15,9 @@ app = Celery('commerce')
 app.config_from_object('django.conf:settings', namespace='CELERY')
 
 # Load task modules from all registered Django app configs.
-app.autodiscover_tasks(lambda: INSTALLED_APPS)
+app.autodiscover_tasks()
 
+
+@app.task(bind=True)
+def debug_task(self):
+    print('Request: {0!r}'.format(self.request))
